+++ pom.xml
@@ -6,7 +6,7 @@
 
   <groupId>ir.sahab.testcontainers</groupId>
   <artifactId>dns-aware-test-container</artifactId>
-  <version>1.1.0</version>
+  <version>1.2.0</version>
 
   <properties>
     <maven.compiler.source>8</maven.compiler.source>
